WORKING VERSION!  Reweighted mapping values, cut point cloud to not be processsing if th points are behind due to transformation errors.  Also now maps the scans that hit nothing

+++ src/Mapper.hpp
@@ -13,7 +13,7 @@
 struct GridUpdate
 {
 	size_t grid_index;
-	int64_t value;
+	double value;
 };
 
 class Mapper
